Extract some persistence value handlers to a separate package (#105)
