+++ aes-core/src/main/java/com/ijioio/aes/core/persistence/jdbc/value/handler/JdbcPersistenceValueHandler.java
@@ -1,17 +1,19 @@
-package com.ijioio.aes.core.persistence.jdbc;
+package com.ijioio.aes.core.persistence.jdbc.value.handler;
 
 import java.util.Collection;
 import java.util.List;
 
 import com.ijioio.aes.core.TypeReference;
 import com.ijioio.aes.core.persistence.PersistenceException;
+import com.ijioio.aes.core.persistence.jdbc.JdbcPersistenceContext;
+import com.ijioio.aes.core.persistence.jdbc.JdbcPersistenceHandler;
 
 public interface JdbcPersistenceValueHandler<T> {
 
 	public Class<T> getType();
 
 	public List<String> getColumns(JdbcPersistenceContext context, JdbcPersistenceHandler handler, String name,
-			TypeReference<T> type, boolean search);
+			TypeReference<T> type, boolean search) throws PersistenceException;
 
 	public void write(JdbcPersistenceContext context, JdbcPersistenceHandler handler, TypeReference<T> type, T value,
 			boolean search) throws PersistenceException;
